requirements.txt
# ──────────────────────────────────────────────────
#  CRAWL — Dependencies
# ──────────────────────────────────────────────────

# Core
playwright>=1.40.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
pandas>=2.1.0

# Async HTTP (for webhooks)
aiohttp>=3.9.0

# Email validation (optional, for MX lookups)
dnspython>=2.4.0

# ── SaaS Backend ──
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
pydantic[email]>=2.5.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6

# Stripe billing
stripe>=8.0.0

# Task queue (optional — degrades gracefully)
celery[redis]>=5.3.0

# Google Sheets integration (optional)
# gspread>=6.0.0
# google-auth>=2.25.0

# Airtable integration (optional)
# pyairtable>=2.2.0
